message building step added
