+++ src/rtgs_lab_tools/device_monitoring/core.py
@@ -2,10 +2,10 @@
 Overview:
     - This file is an implementation of the core functionality of device monitoring
 Task Description:
-    - should organize all the functions from within data_getter, date_formatter, data_analyzer, notification system
+    - should organize all the functions from within data_getter, date_formatter, data_analyzer, message_builder, notification system
     - should create the function that the cli.py wraps
 Work Flow:
-    - data_getter->data_formatter->data_analyzer->notification_system
+    - data_getter->data_formatter->data_analyzer->message_builder->notification_system
 To run:
     - python -m rtgs_lab_tools.device_monitoring.core
 """
@@ -16,6 +16,7 @@ from datetime import datetime, timedelta
 from .data_analyzer import analyze_data
 from .data_formatter import format_data_with_parser
 from .data_getter import get_data
+from .message_builder import build_message
 from .notification_system import notify
 
 
@@ -42,6 +43,11 @@ def monitor(
     analysis_dict = analyze_data(formatted_data)
     print(f"--Data analysis complete--")
 
-    # Step 4: Notify the user with the analysis result
+    # Step 4: Build notification messages
+    print(f"--Beginning message building--")
+    message_dict = build_message(analysis_dict)
+    print(f"--Message building complete--")
+
+    # Step 5: Notify the user with the message
     print("\n--Notification Results--\n")
-    notify(analysis_dict, no_email=no_email)
+    notify(message_dict, no_email=no_email)
